refactor: Move file resolution out of Api; extract ApiFile and DefinitionsFile.readAt
